change cost to double and did csv updaters

+++ src/CSVHandlers/CSVLoader.java
@@ -47,7 +47,7 @@ public class CSVLoader {
                     String studentPhone = parts[3];
                     String ownerOrClientString = parts[4];
                     OWNER_OR_CLIENT ownerOrClient = null;
-                    int credits = Integer.parseInt(parts[5]);
+                    double credits = Integer.parseInt(parts[5]);
                     int vehiclesOwned = Integer.parseInt(parts[6]);
 
                     if (ownerOrClientString.equalsIgnoreCase("owner")){
@@ -241,7 +241,7 @@ public class CSVLoader {
                     int date  = Integer.parseInt(parts[4]);
                     int start = Integer.parseInt(parts[5]);
                     int end   = Integer.parseInt(parts[6]);
-                    int cost  = Integer.parseInt(parts[7]);
+                    double cost  = Integer.parseInt(parts[7]);
 
                     Users student = findUserById(studentId);
                     Vehicles vehicle = getVehicleById(vehicleId);
@@ -289,7 +289,7 @@ public class CSVLoader {
                     LOCATION station= EnumsHandler.getLocation(parts[3]);
                     LocalTime start = LocalTime.parse(parts[4]);
                     LocalTime end   = LocalTime.parse(parts[5]);
-                    int credits     = Integer.parseInt(parts[6]);
+                    double credits     = Double.parseDouble(parts[6]);
                     Users user       = findUserById(studentId);
                     Vehicles vehicle = getVehicleById(vehicleId);
                     if (user != null && vehicle != null) {
